requirements.txt
astroid==2.4.1
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
emoji==0.5.4
future==0.18.2
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imutils==0.5.3
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pycparser==2.20
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==12.7
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.17.2
scipy==1.4.1
six==1.14.0
soupsieve==2.0.1
tifffile==2020.5.11
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
